nemo_guardrails_tracing_project/requirements.txt
pytest>=8.0.0
pyyaml>=6.0.1
openai>=1.0.0
nemoguardrails>=0.9.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0
